Usando context api, reestruturando para components a antiga home, passando funções  e demais .
