+++ aulas/reactjs/aulas/02-Criando-SPAS/02-ignite-timer/src/pages/Home/components/CountDown/index.tsx
@@ -1,0 +1,92 @@
+import { differenceInSeconds } from 'date-fns'
+import { useContext, useEffect } from 'react'
+import { CyclesContext } from '../..'
+import { CountDownContainer, Separator } from './styles'
+
+/* substituido pelo contextApi
+interface CountDownProps {
+  activeCycle: any
+  activeCycleId: any
+  setCycles: any
+} */
+
+/*
+Prop Drilling -> Quanto temos  MUITAS  propriedades APENAS para comunicação entre components
+* Conceito :  CONTEXT API, permite  compartilhar  informações entre components, cria informações "GLOBAIS"
+*/
+
+export function CountDown() {
+  const {
+    activeCycle,
+    activeCycleId,
+    markCurrentCycleAsFinished,
+    amountSecondsPassed,
+    setSecondsPassed,
+  } = useContext(CyclesContext)
+
+  const totalSeconds = activeCycle ? activeCycle.minutesAmount * 60 : 0
+
+  const currentSeconds = activeCycle ? totalSeconds - amountSecondsPassed : 0
+
+  const minutesAmount = Math.floor(currentSeconds / 60)
+  const secondsAmount = currentSeconds % 60
+
+  const minutes = String(minutesAmount).padStart(2, '0')
+  const seconds = String(secondsAmount).padStart(2, '0')
+
+  useEffect(() => {
+    if (activeCycle) {
+      document.title = `${minutes}:${seconds}`
+    }
+  }, [minutes, seconds, activeCycle])
+
+  useEffect(() => {
+    let interval: number
+    if (activeCycle) {
+      interval = setInterval(() => {
+        const secondsDifference = differenceInSeconds(
+          new Date(),
+          activeCycle.startDate,
+        )
+
+        if (secondsDifference > totalSeconds) {
+          // setCycles((state) =>
+          //  state.map((cycle) => {
+          //    if (cycle.id === activeCycleId) {
+          //      return { ...cycle, finishedDate: new Date() }
+          //    } else {
+          //      return cycle
+          //    }
+          //  }),
+          // )
+
+          markCurrentCycleAsFinished()
+
+          clearInterval(interval)
+        } else {
+          setSecondsPassed(secondsDifference)
+        }
+      }, 1000)
+    }
+
+    return () => {
+      clearInterval(interval)
+    }
+  }, [
+    activeCycle,
+    totalSeconds,
+    activeCycleId,
+    markCurrentCycleAsFinished,
+    setSecondsPassed,
+  ])
+
+  return (
+    <CountDownContainer>
+      <span>{minutes[0]}</span>
+      <span>{minutes[1]}</span>
+      <Separator>:</Separator>
+      <span>{seconds[0]}</span>
+      <span>{seconds[1]}</span>
+    </CountDownContainer>
+  )
+}
