+++ aulas/reactjs/aulas/02-Criando-SPAS/02-ignite-timer/src/pages/Home/components/NewCycleForm/styles.ts
@@ -1,0 +1,43 @@
+import styled from 'styled-components'
+
+export const FormContainer = styled.main`
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 0.5rem;
+  color: ${(props) => props.theme['gray-100']};
+  font-size: 1.25rem;
+  font-weight: bold;
+  flex-wrap: wrap;
+`
+
+const BaseInput = styled.input`
+  background: transparent;
+  height: 2.5rem;
+  border: 0;
+  border-bottom: 2px solid ${(props) => props.theme['gray-500']};
+  font-weight: bold;
+  font-size: 1.25rem;
+  padding: 0 0.5rem;
+  color: ${(props) => props.theme['gray-100']};
+
+  &:focus {
+    box-shadow: none;
+    border-color: ${(props) => props.theme['green-500']};
+  }
+
+  &::placeholder {
+    color: ${(props) => props.theme['gray-500']};
+  }
+`
+
+export const TaskInput = styled(BaseInput)`
+  flex: 1;
+  &::-webkit-calendar-picker-indicator {
+    display: none !important;
+  }
+`
+export const MinutesAmountInput = styled(BaseInput)`
+  width: 4rem;
+`
